[36] - Bounded Context creado para DRY
