+++ src/DRY/Principle/Error/Leader.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Deg540\PHPTestingBoilerplate\DRY\Trial;
+namespace Deg540\PHPTestingBoilerplate\DRY\Principle\Error;
 
 class Leader
 {
