+++ src/DRY/Principle/Success/Employee.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Deg540\PHPTestingBoilerplate\DRY\Success;
+namespace Deg540\PHPTestingBoilerplate\DRY\Principle\Success;
 
 abstract class Employee
 {
