+++ src/DRY/Tests/Success/DeveloperSuccessTest.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace Deg540\PHPTestingBoilerplate\Test\DRY\Error;
+namespace Deg540\PHPTestingBoilerplate\DRY\Tests\Success;
 
-use Deg540\PHPTestingBoilerplate\DRY\Error\Developer;
+use Deg540\PHPTestingBoilerplate\DRY\Principle\Success\Developer;
 use PHPUnit\Framework\TestCase;
 
-class DeveloperErrorTest extends TestCase
+class DeveloperSuccessTest extends TestCase
 {
     public Developer $etxebe;
     public Developer $asier;
@@ -53,8 +53,8 @@ class DeveloperErrorTest extends TestCase
      */
     public function noTasksToDevelop()
     {
-        $tasksEtxebe = $this->etxebe->tasksToDevelop();
-        $tasksAsier = $this->asier->tasksToDevelop();
+        $tasksEtxebe = $this->etxebe->tasksToBeProgrammed();
+        $tasksAsier = $this->asier->tasksToBeProgrammed();
 
         $this->assertEquals('You have no tasks to develop', $tasksEtxebe);
         $this->assertEquals('You have tasks to develop', $tasksAsier);
